backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
google-generativeai>=0.2.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.11.1
slowapi>=0.1.9
httpx>=0.25.0
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0



